Extracted CellImage from Cell.cs
